Refactor current species extraction code
